dom/media/mediasource/test/test_NoAudioLoopBackData.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: loop-back data not available yet (shorter audio)</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<script class="testbody" type="text/javascript">

/**
 * This test is used to check whether a looping video can loop back successfully
 * when it has a shorter audio track than its video track. When reaching EOS for
 * the shorter track, there is no loop-back data at the start position (they are
 * not appended yet) Even that, we should still be able to loop back but the
 * looping would become non-seamless in this situation.
 */
SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, el) => {
  await once(ms, "sourceopen");
  ok(true, "Receive a sourceopen event");
  const videosb = ms.addSourceBuffer("video/mp4");
  const audiosb = ms.addSourceBuffer("audio/mp4");

  // Here we create a shorter audio than video.
  info(`create different length source buffers`);
  await fetchAndLoad(videosb, "bipbop/bipbop_video", ["init"], ".mp4");
  await fetchAndLoad(videosb, "bipbop/bipbop_video", range(5, 8), ".m4s");
  audiosb.appendWindowEnd = videosb.buffered.end(0) - 0.2;
  await fetchAndLoad(audiosb, "bipbop/bipbop_audio", ["init"], ".mp4");
  await fetchAndLoad(audiosb, "bipbop/bipbop_audio", range(5, 8), ".m4s");
  ms.endOfStream();
  await Promise.all([once(el, "durationchange"), once(ms, "sourceended")]);
  info(`audio=[${audiosb.buffered.start(0)}-${audiosb.buffered.end(0)}], video=[${videosb.buffered.start(0)}-${videosb.buffered.end(0)}]`);
  ok(true, `endOfStream completed, buffer=[${el.buffered.start(0)}, ${el.buffered.end(0)}]`);
  ok(videosb.buffered.end(0) > audiosb.buffered.end(0), `video should be longer than audio`);

  info(`seek to the position where buffered data exists`);
  el.loop = true;
  el.controls = true;
  el.currentTime = el.buffered.start(0);
  await el.play();

  info(`video should be able to loop back to the start position`);
  await once(el, "seeking");
  await once(el, "seeked");

  SimpleTest.finish(SimpleTest);
});

</script>
</body>
</html>
